requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
spotipy>=2.23.0
python-dotenv>=1.0.0
requests>=2.31.0
PySide6>=6.6.0
